Make ExampleCardList component and refactor inputs page
